yatbc/requirements.txt
django
django-bootstrap5
torbox_api
django-tasks
transmission-rpc
requests
asttokens
django-constance
bleach
dotenv
